requirements.txt
click==6.7
Flask==0.12.2
Jinja2==2.9.6
SQLAlchemy==1.1.14
Flask-SQLAlchemy==2.2
Flask-Alembic
pymysql
Fabric3==1.13.1.post1
pyjwt
gunicorn
